+++ lib/shared/link_preview_card.dart
@@ -1,8 +1,8 @@
-import 'package:flex_color_scheme/flex_color_scheme.dart';
 import 'package:flutter/material.dart';
 
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:url_launcher/url_launcher.dart';
+import 'package:flutter_custom_tabs/flutter_custom_tabs.dart';
+import 'package:url_launcher/url_launcher.dart' hide launch;
 
 import 'package:thunder/user/bloc/user_bloc.dart';
 import 'package:thunder/community/bloc/community_bloc.dart';
@@ -10,8 +10,6 @@ import 'package:thunder/account/bloc/account_bloc.dart';
 import 'package:thunder/community/pages/community_page.dart';
 import 'package:thunder/core/auth/bloc/auth_bloc.dart';
 import 'package:thunder/core/enums/view_mode.dart';
-import 'package:thunder/core/theme/bloc/theme_bloc.dart';
-import 'package:thunder/shared/webview.dart';
 import 'package:thunder/thunder/bloc/thunder_bloc.dart';
 import 'package:thunder/utils/instance.dart';
 import 'package:thunder/shared/image_preview.dart';
@@ -42,7 +40,6 @@ class LinkPreviewCard extends StatelessWidget {
 
   final bool showLinkPreviews;
   final bool showFullHeightImages;
-
 
   final bool edgeToEdgeImages;
 
@@ -98,9 +95,9 @@ class LinkPreviewCard extends StatelessWidget {
           child: inkWell,
         );
       } else {
-          return Padding(
-            padding: const EdgeInsets.only(top: 4.0, bottom: 8.0),
-            child: inkWell,
+        return Padding(
+          padding: const EdgeInsets.only(top: 4.0, bottom: 8.0),
+          child: inkWell,
         );
       }
     }
@@ -114,7 +111,7 @@ class LinkPreviewCard extends StatelessWidget {
       try {
         UserBloc userBloc = BlocProvider.of<UserBloc>(context);
         userBloc.add(MarkUserPostAsReadEvent(postId: postId!, read: true));
-      } catch(e){
+      } catch (e) {
         CommunityBloc communityBloc = BlocProvider.of<CommunityBloc>(context);
         communityBloc.add(MarkPostAsReadEvent(postId: postId!, read: true));
       }
@@ -144,21 +141,38 @@ class LinkPreviewCard extends StatelessWidget {
       if (openInExternalBrowser) {
         launchUrl(Uri.parse(originURL!), mode: LaunchMode.externalApplication);
       } else {
-        Navigator.of(context).push(MaterialPageRoute(builder: (context) => WebView(url: originURL!)));
+        launch(
+          originURL!,
+          customTabsOption: CustomTabsOption(
+            toolbarColor: Theme.of(context).canvasColor,
+            enableUrlBarHiding: true,
+            showPageTitle: true,
+            enableDefaultShare: true,
+            enableInstantApps: true,
+          ),
+          safariVCOption: SafariViewControllerOption(
+            preferredBarTintColor: Theme.of(context).canvasColor,
+            preferredControlTintColor: Theme.of(context).textTheme.titleLarge?.color ?? Theme.of(context).primaryColor,
+            barCollapsingEnabled: true,
+          ),
+        );
       }
     }
   }
 
   Widget linkInformation(BuildContext context) {
     final theme = Theme.of(context);
-    final bool useDarkTheme = context.read<ThemeBloc>().state.useDarkTheme;
 
     if (viewMode == ViewMode.compact) {
       return Container(
         clipBehavior: Clip.hardEdge,
         decoration: BoxDecoration(borderRadius: BorderRadius.circular(6)),
         child: Container(
-          color: useDarkTheme ? theme.colorScheme.background.lighten(7) : theme.colorScheme.background.darken(7),
+          color: ElevationOverlay.applySurfaceTint(
+            Theme.of(context).colorScheme.surface,
+            Theme.of(context).colorScheme.surfaceTint,
+            10,
+          ),
           child: SizedBox(
             height: 75.0,
             width: 75.0,
@@ -171,7 +185,11 @@ class LinkPreviewCard extends StatelessWidget {
       );
     } else {
       return Container(
-        color: useDarkTheme ? theme.colorScheme.background.lighten(7) : theme.colorScheme.background.darken(7),
+        color: ElevationOverlay.applySurfaceTint(
+          Theme.of(context).colorScheme.surface,
+          Theme.of(context).colorScheme.surfaceTint,
+          10,
+        ),
         padding: const EdgeInsets.symmetric(horizontal: 8.0, vertical: 12.0),
         child: Row(
           children: [