+++ lib/thunder/bloc/thunder_state.dart
@@ -7,7 +7,6 @@ class ThunderState extends Equatable {
     this.status = ThunderStatus.initial,
 
     // General
-    this.database,
     this.version,
     this.errorMessage,
 
@@ -20,7 +19,6 @@ class ThunderState extends Equatable {
 
     // Post Settings
     this.collapseParentCommentOnGesture = true,
-    this.disableSwipeActionsOnPost = false,
     this.showThumbnailPreviewOnRight = false,
     this.showLinkPreviews = true,
     this.showVoteActions = true,
@@ -59,9 +57,7 @@ class ThunderState extends Equatable {
     this.rightSecondaryCommentGesture = SwipeAction.save,
 
     // Theme Settings
-    this.useSystemTheme = false,
-    this.themeType = 'dark',
-    this.useBlackTheme = false,
+    this.themeType = ThemeType.system,
     this.useMaterialYouTheme = false,
 
     // Font Scale
@@ -71,7 +67,6 @@ class ThunderState extends Equatable {
 
   final ThunderStatus status;
 
-  final Database? database;
   final Version? version;
 
   final String? errorMessage;
@@ -86,7 +81,6 @@ class ThunderState extends Equatable {
 
   // Post Settings
   final bool collapseParentCommentOnGesture;
-  final bool disableSwipeActionsOnPost;
   final bool showThumbnailPreviewOnRight;
   final bool showLinkPreviews;
   final bool showVoteActions;
@@ -125,9 +119,7 @@ class ThunderState extends Equatable {
   final SwipeAction rightSecondaryCommentGesture;
 
   // Theme Settings
-  final bool useSystemTheme;
-  final String themeType;
-  final bool useBlackTheme;
+  final ThemeType themeType;
   final bool useMaterialYouTheme;
 
   // Font Scale
@@ -136,7 +128,6 @@ class ThunderState extends Equatable {
 
   ThunderState copyWith({
     ThunderStatus? status,
-    Database? database,
     Version? version,
     String? errorMessage,
 
@@ -149,7 +140,6 @@ class ThunderState extends Equatable {
 
     // Post Settings
     bool? collapseParentCommentOnGesture,
-    bool? disableSwipeActionsOnPost,
     bool? showThumbnailPreviewOnRight,
     bool? showLinkPreviews,
     bool? showVoteActions,
@@ -188,9 +178,7 @@ class ThunderState extends Equatable {
     SwipeAction? rightSecondaryCommentGesture,
 
     // Theme Settings
-    bool? useSystemTheme,
-    String? themeType,
-    bool? useBlackTheme,
+    ThemeType? themeType,
     bool? useMaterialYouTheme,
 
     // Font Scale
@@ -199,7 +187,6 @@ class ThunderState extends Equatable {
   }) {
     return ThunderState(
       status: status ?? this.status,
-      database: database ?? this.database,
       version: version ?? this.version,
       errorMessage: errorMessage,
       // Feed Settings
@@ -210,7 +197,6 @@ class ThunderState extends Equatable {
       defaultCommentSortType: defaultCommentSortType ?? this.defaultCommentSortType,
       // Post Settings
       collapseParentCommentOnGesture: collapseParentCommentOnGesture ?? this.collapseParentCommentOnGesture,
-      disableSwipeActionsOnPost: disableSwipeActionsOnPost ?? this.disableSwipeActionsOnPost,
       showThumbnailPreviewOnRight: showThumbnailPreviewOnRight ?? this.showThumbnailPreviewOnRight,
       showLinkPreviews: showLinkPreviews ?? this.showLinkPreviews,
       showVoteActions: showVoteActions ?? this.showVoteActions,
@@ -246,9 +232,7 @@ class ThunderState extends Equatable {
       rightSecondaryCommentGesture: rightSecondaryCommentGesture ?? this.rightSecondaryCommentGesture,
 
       // Theme Settings
-      useSystemTheme: useSystemTheme ?? this.useSystemTheme,
       themeType: themeType ?? this.themeType,
-      useBlackTheme: useBlackTheme ?? this.useBlackTheme,
       useMaterialYouTheme: useMaterialYouTheme ?? this.useMaterialYouTheme,
       // Font Scale
       titleFontSizeScale: titleFontSizeScale ?? this.titleFontSizeScale,
@@ -259,7 +243,6 @@ class ThunderState extends Equatable {
   @override
   List<Object?> get props => [
         status,
-        database,
         version,
         errorMessage,
         useCompactView,
@@ -268,7 +251,6 @@ class ThunderState extends Equatable {
         defaultSortType,
         defaultCommentSortType,
         collapseParentCommentOnGesture,
-        disableSwipeActionsOnPost,
         showThumbnailPreviewOnRight,
         showLinkPreviews,
         showVoteActions,
@@ -295,9 +277,7 @@ class ThunderState extends Equatable {
         leftSecondaryCommentGesture,
         rightPrimaryCommentGesture,
         rightSecondaryCommentGesture,
-        useSystemTheme,
         themeType,
-        useBlackTheme,
         useMaterialYouTheme,
         titleFontSizeScale,
         contentFontSizeScale,