+++ CHANGELOG.md
@@ -16,11 +16,18 @@
 - Settings have been re-organized to be more consistent, and to show available options between different views - contribution from @CTalvio
 - Feed view will no longer show full screen error messages
 - Added comment button actions and added an option to toggle comment button actions
+- Added maximum depth to comments. You can now tap on Load more replies to get more replies for a comment thread within a post
+- Various login flow improvements - contribution from @micahmo
+- In app browser is now switched over to use custom tabs - contribution from @micahmo
+- Adjusted theming options to show a modal rather than toggles to reduce confusion - contribution from @coslu
+- Added ability to share post, external link, or media from post share button - contribution from @micahmo
+- Added ability to disable post and comment swipe actions separately
 
 ### Changed
 - Adjusted subscription styling to be more consistent - contribution from @micahmo
 - Removed Sentry error logging
 - Tapping outside of the text field when creating a comment/reply will dismiss the keyboard
+- Adjusted divider and link preview card colours to have better contrast
 
 ### Fixed
 - Fixed issue with styling differences in compact and normal view for community/instance - contribution from @machinaeZER0
@@ -29,6 +36,8 @@
 - Fixed issue where copy/pasting was non-functional on the Search input - contribution from @ajsosa
 - Fixed issue where subscribing to a community from the search page would not refresh sidebar subscriptions - contribution from @micahmo
 - Fixed issue where only some subscriptions would show up on the sidebar
+- Fixed issue where refreshing community when tapping on a link would show local posts rather than the community's posts
+- Fixed issue where you could not log in through email
 
 ## 0.2.1+11 - 2023-07-09
 ### Added